+++ functions.php
@@ -7,77 +7,12 @@
  * @package r3-id-documentation
  */
 
-namespace BU\theme_slug;
-
-// Disabling phpcs check for default responsi constants.
-// phpcs:disable WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedConstantFound
-/**
- * Current theme's domain.
- */
-define( 'RESPONSIVE_CHILD_THEME_DOMAIN', 'r3-id-documentation' );
+namespace BU\r3_id_documentation;
 
 /**
- * Current theme's server path. No trailing /
+ * Constants.
  */
-define( 'RESPONSIVE_CHILD_THEME_DIR', get_stylesheet_directory() );
-
-/**
- * Current theme's web path. No trailing /
- */
-define( 'RESPONSIVE_CHILD_THEME_URL', get_theme_file_uri() );
-
-/**
- * SEO functionality.
- *
- * @link https://developer.bu.edu/webteam/developer/libraryframework/wordpress/plugins/post-details/#:~:text=after%20the%20text.-,SEO,Output%20meta%20elements.,-Coming%20in%20Version
- */
-define( 'BU_SUPPORTS_SEO', false );
-
-/**
- * Current theme's version.
- *
- * This DOES NOT control the version of Responsive Framework.
- *
- * It represents the version of this child theme and is used to version theme
- * assets (style.css, production.js, etc.) when loaded.
- *
- * @link https://semver.org/
- */
-define( 'RESPONSIVE_CHILD_THEME_VERSION', wp_get_theme()->get( 'Version' ) );
-
-/**
- * Specifies the responsive layout for the theme. Disables the option in the
- * Customizer.
- *
- * Removing this allows the layout to be changed in the Customizer.
- *
- * Available values: `default`, `top-nav`, `side-nav`, `no-nav`, `mega-nav`.
- *
- * @link https://github.com/bu-ist/responsive-framework/wiki/Changing-Available-Layouts-And-Default-Layout
- */
-define( 'BU_RESPONSIVE_LAYOUT', 'default' );
-
-/**
- * Specifies the site's branding type. Disables the option in the Customizer.
- *
- * Removing this allows the branding type to be changed in the Customizer.
- *
- * Available values: `logotype`, `signature`, `unbranded`.
- */
-define( 'BU_BRANDING_TYPE', 'logotype' );
-
-/**
- * Defines a sidebar position for the theme. Disables the option in the
- * Customizer.
- *
- * Removing this allows the layout to be changed in the Customizer.
- *
- * Available values: `right`, `left`, `bottom`.
- */
-define( 'BU_RESPONSIVE_SIDEBAR_POSITION', 'right' );
-
-// Re-enabling constant check.
-// phpcs:enable WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedConstantFound
+require_once 'constants.php';
 
 /**
  * @todo: Delete the following code block and the dev/prototyping folder before going live.
@@ -89,43 +24,22 @@ if ( file_exists( __DIR__ . '/dev/prototyping/dev-functions.php' ) ) {
 	add_filter( 'after_setup_theme', 'dev_sections' );
 }
 
-
-
-
-// // Ensure WordPress environment is loaded
-// require_once ABSPATH . 'wp-load.php';
-
-// // Get all attachments (media items)
-// $attachments = get_posts(
-//  array(
-//      'post_type'      => 'attachment',
-//      'posts_per_page' => 100, // Get all attachments
-//      'post_status'    => 'any', // Include attachments regardless of status
-//  )
-// );
-
-// // Loop through each attachment and delete it
-// foreach ( $attachments as $attachment ) {
-//  wp_delete_attachment( $attachment->ID, true ); // true for permanent deletion
-// }
-
-// echo 'All media items deleted.';
-
-
-
 /**
  * Theme Supports.
  */
 require_once 'includes/theme-support.php';
 
-
-// WP Docs menu
-// Example code for additional items:
-	//$wpdocs_menu[] = [
-		//  'id'    => '',
-		//  'title' => __( '', 'r3-id-documentation' ),
-		//  'href'  => '',
-		//];
+/**
+ * WP Docs menu.
+ *
+ * Example code for additional items:
+ *
+ * $wpdocs_menu[] = [
+ *   'id'    => '',
+ *   'title' => __( '', 'r3-id-documentation' ),
+ *   'href'  => '',
+ * ];
+ */
 add_filter(
 	'wpdocs_menu',
 	function ( $wpdocs_menu ) {
@@ -143,9 +57,3 @@ require_once 'includes/class-themeenqueues.php';
  * Blocks & Assets
  */
 require_once 'includes/blocks/functions.php';
-
-// Uncomment to unbundle plugin styles.
-// add_filter( 'plugin_slug_bundle_block_styles', '__return_false' );
-
-// Uncomment to remove decorative plugin styles.
-// add_filter( 'plugin_slug_decorative_block_styles', '__return_false' );