+++ src/blocks/_includes/components/BackgroundControl/index.mjs
@@ -1,1 +1,0 @@
-// React component example that is shared between multiple blocks in this plugin.