+++ webpack.customizations.js
@@ -30,6 +30,7 @@ const customIncludePaths = [
  */
 const themeEntryPoints = {
 	// Styles
+	'css/normalize': './src/scss/normalize.scss', // Build a Normalize stylesheet.
 	'css/theme': './src/scss/theme.scss', // A stylesheet for the theme.
 	'css/admin': './src/scss/admin.scss', // A stylesheet for admin only styles.
 	'css/editor-styles': './src/scss/editor-styles.scss', // A stylesheet for editor only styles.
@@ -51,11 +52,29 @@ const themeEntryPoints = {
 const sassCompiler = 'sass-embedded';
 
 /**
+ * Set sassOptions.
+ * @link https://sass-lang.com/documentation/js-api/interfaces/options/
+ */
+const customSassOptions = {
+	includePaths: customIncludePaths, // Adding our custom include paths.
+	// outputStyle: 'compressed', // Determines the output format of the final CSS style.
+	quietDeps: true, // If this option is set to true, Sass won’t print warnings that are caused by dependencies.
+	silenceDeprecations: [
+		'legacy-js-api',
+		'global-builtin',
+		'import',
+		'slash-div',
+		'color-functions',
+		'color-4-api',
+	],
+};
+
+/**
  * The stats option lets you precisely control what bundle information gets displayed. This can be a nice middle ground if you don't want to use quiet or noInfo because you want some bundle information, but not all of it.
  *
  * It is defined as a const here so it can be used for both blocksConfig and themeConfig, but if you'd like to use a different setup for each, you can define them inside the individual config objects.
  *
- * @link https://webpack.js.org/configuration/stats/
+ * @see https://webpack.js.org/configuration/stats/
  */
 const statsConfig = {
 	preset: 'verbose', // Output everything.
@@ -65,7 +84,7 @@ const statsConfig = {
 	colors: true, // Use colors for better readability.
 	modules: false, // Hide module details.
 	chunks: false, // Hide chunk details.
-	chunkGroups : false,
+	chunkGroups: false,
 	assets: false, // Hide "assets by path" details.
 	entrypoints: true, // Hide the entry points with the corresponding bundles.
 	logging: 'none',
@@ -86,8 +105,8 @@ const statsConfig = {
  * Export these so webpack.config.js can consume it.
  */
 module.exports = {
-	customIncludePaths,
 	themeEntryPoints,
 	sassCompiler,
 	statsConfig,
+	customSassOptions,
 };