+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 - Adds `<Image>` sandbox block
 
+## FORTHCOMING
+
+## 0.3.0
+
+- Added Embed Markdown block
+- Fixed block category bug
+
+## 0.2.0
+
+- Replaced all core theme files with latest `responsive-child-starter-3x-block-editor`
+
 ## 0.1.0
 
 - Adds `<LoadingSpinner>` sandbox block